+++ src/preprocessing/passes/synth_rew_rules.cpp
@@ -395,8 +395,7 @@ PreprocessingPassResult SynthRewRulesPass::applyInternal(
   {
     datatypes.push_back(sdts[i].getDatatype());
   }
-  std::vector<TypeNode> types = nm->mkMutualDatatypeTypes(
-      datatypes, NodeManager::DATATYPE_FLAG_PLACEHOLDER);
+  std::vector<TypeNode> types = nm->mkMutualDatatypeTypes(datatypes);
   Trace("srs-input") << "...finished." << std::endl;
   Assert(types.size() == datatypes.size());
   std::map<Node, TypeNode> subtermTypes;
@@ -440,8 +439,7 @@ PreprocessingPassResult SynthRewRulesPass::applyInternal(
     // set that this is a sygus datatype
     sdttl.initializeDatatype(t, sygusVarList, false, false);
     DType dttl = sdttl.getDatatype();
-    TypeNode tlt =
-        nm->mkDatatypeType(dttl, NodeManager::DATATYPE_FLAG_PLACEHOLDER);
+    TypeNode tlt = nm->mkDatatypeType(dttl);
     tlGrammarTypes[t] = tlt;
     Trace("srs-input") << "Grammar is: " << std::endl;
     Trace("srs-input") << printer::smt2::Smt2Printer::sygusGrammarString(tlt)